requirements.txt
flask
python-dotenv
SQLAlchemy
flask-sqlalchemy
pymysql